# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Meridian API
  description: |
    Marketing Analytics & Causal Inference Platform
    
    Meridian provides APIs for:
    - **Uplift Modeling**: Predict treatment effects (CATE) for customers
    - **Demand Forecasting**: Generate probabilistic demand forecasts
    - **Price Optimization**: Calculate optimal prices based on elasticity
    - **A/B Experiments**: Manage and analyze experiments
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary

servers:
  - url: https://api.meridian.example.com
    description: Production
  - url: https://api-staging.meridian.example.com
    description: Staging
  - url: http://localhost:8000
    description: Local development

security:
  - bearerAuth: []
  - apiKey: []

tags:
  - name: health
    description: Health check endpoints
  - name: uplift
    description: Uplift prediction endpoints
  - name: forecasting
    description: Demand forecasting endpoints
  - name: pricing
    description: Price optimization endpoints
  - name: experiments
    description: A/B experiment management

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/uplift/predict:
    post:
      tags: [uplift]
      summary: Predict uplift for customers
      description: Returns Conditional Average Treatment Effect (CATE) predictions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpliftPredictionRequest'
      responses:
        '200':
          description: Successful predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpliftPredictionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  /v1/forecast/demand:
    post:
      tags: [forecasting]
      summary: Forecast demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemandForecastRequest'
      responses:
        '200':
          description: Successful forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandForecastResponse'

  /v1/experiments:
    get:
      tags: [experiments]
      summary: List experiments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, running, completed, archived]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of experiments
    post:
      tags: [experiments]
      summary: Create experiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperimentRequest'
      responses:
        '201':
          description: Experiment created

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: healthy
        version:
          type: string
          example: 1.0.0
        timestamp:
          type: string
          format: date-time

    UpliftPredictionRequest:
      type: object
      required:
        - customer_ids
      properties:
        customer_ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 1000
        model_id:
          type: string
          default: causal_forest_v1
        return_confidence_intervals:
          type: boolean
          default: false

    UpliftPredictionResponse:
      type: object
      properties:
        predictions:
          type: array
          items:
            $ref: '#/components/schemas/UpliftPrediction'
        model_id:
          type: string
        model_version:
          type: string

    UpliftPrediction:
      type: object
      properties:
        customer_id:
          type: string
        cate:
          type: number
          description: Conditional Average Treatment Effect
        confidence_interval:
          type: object
          properties:
            lower:
              type: number
            upper:
              type: number

    DemandForecastRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              item_id:
                type: string
              store_id:
                type: string
        horizon_days:
          type: integer
          default: 14

    DemandForecastResponse:
      type: object
      properties:
        forecasts:
          type: array
          items:
            type: object

    CreateExperimentRequest:
      type: object
      required:
        - name
        - hypothesis
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        hypothesis:
          type: string
          minLength: 10
        variants:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              weight:
                type: number

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'